Teleop and IR Range
This is a big commit. I made a working teleop for the dummy box that
will most likely be expanded for the final robot. I also wrote a program
that beeps when the IR sensor isn't getting a signal to do the vertical
range experiment. A few tweeks were made to the low goal autonomous. The
sample teleop is for the younger team to reference